# Ix: Improve Defer() by removing indirection